validaciones de delete

+++ TP2L02/TP2/UI.Web/Comisiones.aspx.cs
@@ -189,8 +189,16 @@ namespace UI.Web
             switch (this.FormMode)
             {
                 case FormModes.Baja:
-                    this.DeleteEntity(this.SelectedID);
-                    this.LoadGrid();
+                    if (Business.Logic.ComisionLogic.isDeleteValid(this.SelectedID))
+                    {
+                        this.DeleteEntity(this.SelectedID);
+                        this.LoadGrid();
+                        this.formPanel.Visible = false;
+                    }
+                    else
+                    {
+                        ScriptManager.RegisterClientScriptBlock(this, this.GetType(), "Esta comision no puede ser eliminada", "alert('Esta comision no puede ser eliminada por que ya esta asignada a un curso')", true);
+                    }
                     break;
                 case FormModes.Modificacion:
                     this.Entity = new Comision();
@@ -206,11 +214,11 @@ namespace UI.Web
                     this.LoadEntity(this.Entity);
                     this.SaveEntity(this.Entity);
                     this.LoadGrid();
+                    this.formPanel.Visible = false;
                     break;
                 default:
                     break;
             }
-            this.formPanel.Visible = false;
         }
         private void EnableForm(bool check)
         {
